Feature/8 userlocation (#53)
* feat(UserLocation): implement userlocation get command

* wip(userlocations): start implementing user locations

* feat: implement adding and deleting user locations

* fix: correct errors in userlocation code and refactor

* chore(release): 1.7.0 [skip ci]

# [1.7.0](https://github.com/GlobalX/PockyBot.NET/compare/v1.6.2...v1.7.0) (2020-03-03)

### Features

* add keywords trigger ([b9e6bbd](https://github.com/GlobalX/PockyBot.NET/commit/b9e6bbd3117ca594a3cd52fc67555164ff9e5e9e))

* test: try testing stuff

* test: fix tests

* test: test userlocation responder

* refactor: userlocationservice into separate services for each operation

* test(user-location-getter): test basic error case

* test(user-location-getter): test the basic cases

* fix(user-location-getter): handle mentioned users without locations

* fix(user-location-deleter): correct logic around when to delete whom

* fix(user-location-setter): improve permissions handling

* feat(help): add userlocation config to help

* chore: upgrade to dotnet 3.1

* chore: fix code style

* fix: undo merge error

* refactor: remove unused code

* fix: do null check before deleting database objects

* fix: don't add unnecessary properties to objects

+++ src/PockyBot.NET/Persistence/Repositories/LocationRepository.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
 using PockyBot.NET.Persistence.Models;
 
 namespace PockyBot.NET.Persistence.Repositories
@@ -31,9 +32,13 @@ namespace PockyBot.NET.Persistence.Repositories
 
         public async Task DeleteLocation(string location)
         {
-            var locationToDelete = _context.Locations.FirstOrDefault(x => x.Name == location);
-            _context.Locations.Remove(locationToDelete);
-            await _context.SaveChangesAsync();
+            var locationToDelete = await _context.Locations.FirstOrDefaultAsync(x => x.Name == location);
+
+            if (locationToDelete != null)
+            {
+                _context.Locations.Remove(locationToDelete);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
